moved transformPath to a separate lib module
