+++ src/commands/replace.ts
@@ -1,0 +1,5 @@
+async function replace(pofile: string, paths: string[]) {
+    console.log("replace");
+}
+
+export default replace;
